Added custom exceptions to global exception handler

+++ src/main/java/org/example/applicationtracker/service/ApplicationService.java
@@ -1,5 +1,6 @@
 package org.example.applicationtracker.service;
 
+import org.example.applicationtracker.exception.ApplicationNotFoundException;
 import org.example.applicationtracker.model.Application;
 import org.example.applicationtracker.repository.ApplicationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,12 +23,19 @@ public class ApplicationService {
 
     public Page<Application> getAllApplications(int page,int size){
         Pageable pageable = PageRequest.of(page, size);
-        return applicationRepository.findAll(pageable);
+        if(applicationRepository.count()==0){
+            return null;
+        }
+        else {
+            return applicationRepository.findAll(pageable);}
     }
     public Application addApplication(Application application){
         return applicationRepository.save(application);
     }
     public void deleteApplication(int id){
+        if(!applicationRepository.existsById(id)){
+            throw new ApplicationNotFoundException(id);
+        }
         applicationRepository.deleteById(id);
     }
 
@@ -35,13 +43,23 @@ public class ApplicationService {
         return applicationRepository.findByJobTitleContaining(keyword);
     }
     public List<Application> getByStatus(String status){
+        if(applicationRepository.findByStatus(status).isEmpty()){
+            throw new ApplicationNotFoundException();
+        }
         return applicationRepository.findByStatus(status);
     }
     public List<Application> getByAppliedDate(LocalDate appliedDate){
+        if(applicationRepository.findByAppliedDate(appliedDate).isEmpty()){
+            throw new ApplicationNotFoundException();
+        }
         return applicationRepository.findByAppliedDate(appliedDate);
     }
     public List<Application> getByCompanyNameAndAppliedDate(String companyName,LocalDate appliedDate){
+        if(applicationRepository.findByCompanyNameAndAppliedDate(companyName, appliedDate).isEmpty()){
+            throw new ApplicationNotFoundException();
+        }
         return applicationRepository.findByCompanyNameAndAppliedDate(companyName,appliedDate);
     }
+
 
 }
